new office fixture, simplification for parent / child office picking; office adding possible now

+++ app/models/office.php
@@ -28,7 +28,6 @@ class Office extends AppModel {
 			'required' => array(
 				'rule' => 'notEmpty',
 				'message' => 'Please specify at least one frequency option!',
-				'required' => true,
 				'last' => true
 			)
 		),
@@ -36,7 +35,6 @@ class Office extends AppModel {
 			'required' => array(
 				'rule' => array('validateAmounts'),
 				'message' => 'Please specify valid amount options!',
-				'required' => true,
 				'last' => true
 			),
 		),
@@ -48,18 +46,23 @@ class Office extends AppModel {
  * @access public
  */
 	function beforeSave() {
-		$this->data['Office']['amounts'] = r(' ', '', $this->data['Office']['amounts']);
+		if (isset($this->data['Office']['amounts'])) {
+			$this->data['Office']['amounts'] = r(' ', '', $this->data['Office']['amounts']);
+		}
 
-		$this->GatewaysOffice->deleteAll(array('office_id' => $this->id));
-		if (!empty($this->data['Office']['gateways'][0])) {
-			foreach ($this->data['Office']['gateways'] as $gatewayId) {
-				$this->GatewaysOffice->create(array(
-					'office_id' => $this->id,
-					'gateway_id' => $gatewayId
-				));
-				$this->GatewaysOffice->save();
+		if (isset($this->data['Office']['gateways'])) {
+			$this->GatewaysOffice->deleteAll(array('office_id' => $this->id));
+			if (!empty($this->data['Office']['gateways'][0])) {
+				foreach ($this->data['Office']['gateways'] as $gatewayId) {
+					$this->GatewaysOffice->create(array(
+						'office_id' => $this->id,
+						'gateway_id' => $gatewayId
+					));
+					$this->GatewaysOffice->save();
+				}
 			}
 		}
+
 		return true;
 	}
 /**
@@ -119,10 +122,7 @@ class Office extends AppModel {
  */
 	function parentOfficeOptions($id) {
 		return $this->find('list', array(
-			'conditions' => array(
-				'id <>' => $id,
-				'parent_id' => ''
-			),
+			'conditions' => array('id <>' => $id),
 			'contain' => false
 		));
 	}
@@ -134,11 +134,14 @@ class Office extends AppModel {
  * @access public
  */
 	function subOfficeOptions($id, $type = 'normal') {
+		$conditions = array(
+			'id <>' => $id
+		);
+		if ($type == 'selected') {
+			$conditions['parent_id'] = $id;
+		}
 		$subOffices = $this->find('list', array(
-			'conditions' => array(
-				'id <>' => $id,
-				'parent_id' => array('', $id)
-			),
+			'conditions' => $conditions,
 			'contain' => false
 		));
 
@@ -150,21 +153,8 @@ class Office extends AppModel {
 			return $ids;
 		}
 
-		$subParentOffices = $this->find('all', array(
-			'conditions' => array(
-				'parent_id' => $ids
-			),
-			'contain' => false,
-			'fields' => array('parent_id')
-		));
-		$ids = Set::extract('/Office/parent_id', $subParentOffices);
-
-		foreach ($subOffices as $id => $name) {
-			if (in_array($id, $ids)) {
-				unset($subOffices[$id]);
-			}
-		}
-
+		// @todo this is simplified; could do MPTT-Pattern here with the tree behavior
+		// however, since only roots edit the office tree, we simplify this
 		return $subOffices;
 	}
 /**
@@ -205,6 +195,28 @@ class Office extends AppModel {
 		$Session = Common::getComponent('Session');
 		if ($id == $Session->read('Office.id')) {
 			$this->activate($id);
+		}
+	}
+/**
+ * undocumented function
+ *
+ * @param string $id 
+ * @return void
+ * @access public
+ */
+	public function subRelations($id, $data) {
+		$this->updateAll(
+			array('Office.parent_id' => "''"),
+			array('Office.parent_id' => $id)
+		);
+
+		if (!is_array($data['Office']['suboffice_id'])) {
+			return false;
+		}
+
+		foreach ($data['Office']['suboffice_id'] as $subOfficeId) {
+			$this->id = $subOfficeId;
+			$this->saveField('parent_id', $id);
 		}
 	}
 }
